MDL-63070 core_form: Move lib changes into traits
MDL-30168 changed the way element ids were generated, particularly for
repeated elements, and should have been made in the MoodleForm wrappers,
not in the QuickForm library.
This change moves this code into traits which are used in the element
wrappers and modifies the formslib code dealing with repeated elements,
making sure that all copied elements have their IDs regenerated
properly.