MDL-68698 tool_moodlenet: update registry to support url type
Part of MDL-67883.
In addition to 'files', allow the registry to capture registrations
for the 'url' type. These are stored in a new array representing all
handler registrations in the registry. As per dnd code, files and
other types will be kept separate, which we achieve via indexes in
the registry array.