MDL-68698: added name and description support
Present the name during the workflow steps and when saving the mod.
Description is not shown during the workflow, but it set after creating
the mod if either of the following is true:
a) intro wasn't set by the mod during creation (i.e. the field is empty)
b) the mod is one of an approved list of mods supporting post-setting of
the description field. Only those mods which do not require description
as part of their display are able to have their descriptions overridden.