Deep Linking: add deep linking support to enrol_lti
Things done:
-------------
- added private key and kid during save of enrol methods and for
existing instances via an upgrade step.
- minor fix to IMS lib to conditionally set 'data' in Deep Link response
only if it's provided in the deep link request.
- modified the enrol/lti/tool.php code to handle deep link launches.
This is currently very hacky and specific.
- Added mod/tooltest/configure.php which handles the results of the user
selection on the enrol/lti/tool.php page (when handling deep link
launches) by returning a deep_link_resource including the relevant tool
difficulty.
Changed the mod/tooltest/view.php page so that it accepts a difficulty
param and loads bombs instead of stars if set to 'hard'. This makes the
subsequent levels harder as the rogue bomb blends in with the
collectible bombs.
TODO:
-----
- tool.php must be generic and shouldn't provide a tooltest specific
content item list. It should delegate to the tool plugin. Maybe this
can be implemented via a callback which plugins implement to direct
the lti flow to their relevant deep link settings page.
- Check what 'data' value is sent in the deep linking request with the
IMS validator tool. If that's a CSRF token, do we need one? I note the
spec's example has a CSRF token in it.