Very crude example of login prompt + email confirmation
This adds the ability to select an account when launching, if not bound
already. Users can login, via Moodle login, and are then asked to
confirm the account link via email.

Problems:
1. Clicking the link in the confirmation email will result in a redirect
to the launch url, which doesn't really work when the original launch
was an embedded launch.
2. When the confirmation email is sent, the user cannot do anything more
in that launch view. They must confirm and reload, or just follow the
redirect currently present in the confirmation step.
3. If you refresh after the confirmation email has been sent, that page
currently does not prevent unconfirmed account access - we need to add
a message there in case that page is reloaded prior to confirmation

Alternatives:
1. Require user session before logins can be linked. This way, we force
the user to authenticate with the tool first, and then the subsequent
account linking process can take them right to the deep link picker or
launch content. No login/index wantsurl dependency. No email confirm.
No funny redirects breaking embedded launches when confirming. Student
linking this way may be tricky, so we may need a way to disable the auto
provisioning fallback in 4.1 work. Consider email, firstname, surname
matching as a first preference too, only asking for the user to auth
with the tool as a means to avoid creating a new account entirely.