MDL-XXXXX enrol_lti: db changes facilitating draft registrations
Summary:
- Allows nulls for a number of fields meaning largely empty drafts can
be stored.
- Adds 'localclientid', an id unique to each issuer which is intended
to be used as a parameter in the initiate_login_uri provided to
platforms. Platforms will thus include this on initiate login requests.
Moodle will be able to uniquely identify the registration based on a
combination of iss and localclientid in the absence of the optional
'client_id' claim in the LTI Advantage spec.
- Add 'status', used to denote draft or active registrations.