Simple test activity packaging
This plugs a simple backup procedure into the 'Share with Moodlenet'
action link, mainly as a means to test the use of non-interactive
backups and their inputs/outputs.

The code serves as proof of the following:
- Changing a non-interactive backup plan's settings. This allows
the backup controller for a non-interactive backup to behave the same
as an interactive backup which supports user-selected settings. The same
settings can be overridden by accessing task settings.
- Storing a backup in a plugin specific file area. i.e. the backup
isn't visible on course or activity backup pages so it won't be mixed
up with regular backups.

Note: tool_moodlenet doesn't implement tool_moodlenet_pluginfile, so
this just uses send_stored_file to directly send the backup when the
'download' optional param is set. This is just a quick hack to get
access to the generated backup file for comparison/testing purposes.