MDL-76588 core_grades: Refactor calculate_average for hidden grades case.
On grader report we include hidden grades in averages calculations.
On user report we don't include hidden grades. So to allow using of
calculate_averages it needs to be refactored for this case.

We also have to add group id param to calculate_average since for
grader report we calculate both overall average and group average