MDL-78576 mod_lti: move count subselect into report base fields
This gets around a limitation of the RB; its inability to use a column
subselect value in another column. This change permits the use of the
count in more than just the 'usage' column, as the row_callback is able
to set an instance var, usable by other column callbacks.