MDL-79843 core: add support for plugin and plugintype deprecation
This allows for plugins to be deprecated much in the same way as they
can be marked as deleted, via the plugin manager methods. Also supports
plugin types. Deprecated plugins are displayed as such during upgrade,
but will not block the upgrade process, nor will they interfere with
existing plugin statuses.