chasing down some other behat fails
This commit highlights that, after fixing the bug in core_completion,
we now have the incomplete state being reported for the "require passing
grade" criteria when the student hasn't received a passing grade, meaning
the UI will change to report this as 'todo' now (the indicator for
incomplete) instead of 'complete - fail grade' (the indicator for
completion_complete_fail, which was set in error until now).

There will very likely be many, many other tests like this example.